Change some "ifs" to "switch case" in playing state

+++ Source/Util/Noise_Generator.h
@@ -27,7 +27,7 @@ namespace Noise
     class Generator
     {
         public:
-            int getValue(int x, int z, int tileX, int tileZ) const;
+            double getValue(int x, int z, int tileX, int tileZ) const;
 
             void setNoiseFunction(const Noise::Data& data);
